refactor: segmented the AI4E_CROP_DATA plugin in AI4E_CROP_DATA and AI4E_LABEL_DATA

+++ business/training/minio_upload_dataset.py
@@ -5,12 +5,16 @@ __VER__ = '0.1.0.0'
 _CONFIG = {
   **BasePlugin.CONFIG,
 
-  'ALLOW_EMPTY_INPUTS' : True,
+  'ALLOW_EMPTY_INPUTS': True,
 
-  'DATASET_OBJECT_NAME' : None,
-  'DATASET_LOCAL_PATH'  : None,
+  'DATASET_OBJECT_NAME': None,
+  'DATASET_LOCAL_PATH': None,
+  'DELETE_AFTER_ZIP': False,
+  'INCLUDE_DIR_IN_ZIP': False,
 
-  'PLUGIN_LOOP_RESOLUTION' : 1/5,
+  'IS_RAW': False,
+
+  'PLUGIN_LOOP_RESOLUTION': 1/10,
 
 
   'VALIDATION_RULES': {
@@ -19,11 +23,13 @@ _CONFIG = {
 
 }
 
+
 class MinioUploadDatasetPlugin(BasePlugin):
   CONFIG = _CONFIG
-  def __init__(self, **kwargs):
+
+  def on_init(self):
+    super(MinioUploadDatasetPlugin, self).on_init()
     self.__uploaded = False
-    super(MinioUploadDatasetPlugin, self).__init__(**kwargs)
     return
 
   def startup(self):
@@ -32,26 +38,36 @@ class MinioUploadDatasetPlugin(BasePlugin):
     assert self.cfg_dataset_local_path is not None
     return
 
-  @property
-  def cfg_dataset_object_name(self):
-    return self._instance_config['DATASET_OBJECT_NAME']
-
-  @property
-  def cfg_dataset_local_path(self):
-    return self._instance_config['DATASET_LOCAL_PATH']
-
   def _process(self):
     if not self.__uploaded:
+      local_path = self.cfg_dataset_local_path
+      # In case the local path is a directory, zip it if it is not already zipped
+      if self.os_path.isdir(local_path) and not self.os_path.exists(f'{local_path}.zip'):
+        zip_path = self.diskapi_zip_dir(local_path, include_dir=self.cfg_include_dir_in_zip)
+        if self.cfg_delete_after_zip:
+          self.diskapi_delete_directory(local_path)
+        local_path = zip_path
+      # endif local_path is a directory and not zipped
+      # In case the local path is a directory and the zip file exists, use the zip file
+      local_path = local_path if local_path.endswith('.zip') else f'{local_path}.zip'
       target_path = self.cfg_dataset_object_name
-      if not target_path.endswith('.zip'):
-        target_path += '.zip'
-      self.upload_file(
-        file_path=self.cfg_dataset_local_path,
-        target_path=target_path,
-        force_upload=True,
+      target_path = target_path if target_path.endswith('.zip') else f'{target_path}.zip'
+      if self.os_path.exists(local_path):
+        self.upload_file(
+          file_path=local_path,
+          target_path=target_path,
+          force_upload=True,
+        )
+        self.__uploaded = True
+        self.cmdapi_stop_current_stream()
+      # endif local path exists
+    # endif dataset uploaded
+    # Report upload status if final dataset
+    if not self.cfg_is_raw and self.__uploaded:
+      self.add_payload_by_fields(
+        is_final_dataset_status=not self.cfg_is_raw,
+        uploaded=self.__uploaded
       )
-      self.__uploaded = True
-      self.cmdapi_stop_current_stream()
-    #endif
+    # endif report
 
     return
